refactored, added the EndOfDay logic

+++ DatScheduleServer.Model/Task.cs
@@ -2,7 +2,7 @@
 {
     public class Task
     {
-        public Task(string name, double duration, TaskType type, string colorCode)
+        public Task(string name, int duration, TaskType type, string colorCode)
         {
             Name = name;
             Duration = duration;
@@ -10,7 +10,7 @@
             ColorCode = colorCode;
         }
 
-        public double Duration { get; private set; }
+        public int Duration { get; private set; }
         public string Name { get; private set; }
         public string ColorCode { get; private set; }
         public TaskType Type { get; private set; }
